requirements.txt
Flask==3.0.3
Flask-Session==0.5.0
Pillow==12.0.0
